+++ backend/src/main/java/ssafy/a303/backend/property/dto/response/PropertyMapDto.java
@@ -3,8 +3,8 @@ package ssafy.a303.backend.property.dto.response;
 public record PropertyMapDto(
         String address,
         String propertyNm,
-        Integer latitude,
-        Integer longitude,
+        Double latitude,
+        Double longitude,
         Double area,
         Integer areaP,
         Long deposit,
